# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /adventurers:
    get:
      operationId: AdventurersController_findAll
      parameters:
        - name: dailyRateOrder
          required: false
          in: query
          description: Tri par taux journalier
          schema:
            enum:
              - asc
              - desc
            type: string
        - name: xpMax
          required: false
          in: query
          description: Expérience maximale (incluse)
          schema:
            example: 50
            type: number
        - name: xpMin
          required: false
          in: query
          description: Expérience minimale (incluse)
          schema:
            example: 10
            type: number
        - name: specialityId
          required: false
          in: query
          description: Filtrer par identifiant de spécialité
          schema:
            example: 3
            type: number
        - name: name
          required: false
          in: query
          description: Filtrer par nom (contains, insensible à la casse)
          schema:
            example: aria
            type: string
      responses:
        '200':
          description: List of adventurers (avec filtres et tri)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Aria Stormblade
                    level:
                      type: number
                      example: 12
                    specialityId:
                      type: number
                      example: 3
                    statusId:
                      type: number
                      example: 1
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:34:56.000Z'
      tags:
        - Adventurers
    post:
      operationId: AdventurersController_create
      parameters: []
      requestBody:
        required: true
        description: New adventurer payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Aria Stormblade
                specialityId:
                  type: number
                  example: 3
                level:
                  type: number
                  example: 1
                statusId:
                  type: number
                  example: 1
              required:
                - name
                - specialityId
      responses:
        '201':
          description: Adventurer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 42
                  name:
                    type: string
                    example: Aria Stormblade
                  level:
                    type: number
                    example: 1
                  specialityId:
                    type: number
                    example: 3
                  statusId:
                    type: number
                    example: 1
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
      tags:
        - Adventurers
  /adventurers/{id}:
    get:
      operationId: AdventurersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Adventurer ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Adventurer by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Aria Stormblade
                  level:
                    type: number
                    example: 12
                  specialityId:
                    type: number
                    example: 3
                  statusId:
                    type: number
                    example: 1
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:34:56.000Z'
      tags:
        - Adventurers
    patch:
      operationId: AdventurersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Adventurer ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Aria Nightwind
                specialityId:
                  type: number
                  example: 2
                level:
                  type: number
                  example: 13
                statusId:
                  type: number
                  example: 2
              additionalProperties: false
      responses:
        '200':
          description: Updated adventurer
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 42
                  name:
                    type: string
                    example: Aria Nightwind
                  level:
                    type: number
                    example: 13
                  specialityId:
                    type: number
                    example: 2
                  statusId:
                    type: number
                    example: 2
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:45:00.000Z'
      tags:
        - Adventurers
...
  /consumables/{id}:
    get:
      operationId: ConsumablesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Consumables
    patch:
      operationId: ConsumablesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConsumableDto'
      responses:
        '200':
          description: ''
      tags:
        - Consumables
    delete:
      operationId: ConsumablesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - Consumables
...
  /consumable-types:
    post:
      operationId: ConsumableTypesController_create
      parameters: []
      requestBody:
        required: true
        description: New consumable type payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Health Potion
              required:
                - name
              additionalProperties: false
      responses:
        '201':
          description: Consumable type created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Health Potion
      security:
        - bearer: []
      tags:
        - Consumable Types
    get:
      operationId: ConsumableTypesController_findAll
      parameters: []
      responses:
        '200':
          description: List of consumable types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Health Potion
      tags:
        - Consumable Types
  /consumable-types/{id}:
    get:
      operationId: ConsumableTypesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable type ID
          schema:
            example: 5
            type: number
      responses:
        '200':
          description: Consumable type by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Health Potion
      tags:
        - Consumable Types
    patch:
      operationId: ConsumableTypesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable type ID
          schema:
            example: 5
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Greater Health Potion
              additionalProperties: false
      responses:
        '200':
          description: Updated consumable type
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Greater Health Potion
      security:
        - bearer: []
      tags:
        - Consumable Types
    delete:
      operationId: ConsumableTypesController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Consumable type ID
          schema:
            example: 5
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Consumable Types
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http